create the provider and value directly in the context file
